版本暫放區/requirements_v2.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
qrcode[pil]>=7.4
psycopg2-binary>=2.9.9
cloudinary>=1.36.0
